wen-Version Estable

+++ td/web/WEB-INF/jsp/modificarModelo.jsp
@@ -9,6 +9,7 @@
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@include file="../../interfaz/inicio.html" %>
+<script type="text/javascript" src="interfaz/verificaciones.js"></script>
 <!DOCTYPE html>
 <div id="content"><div class="inner_copy"><a href="http://www.freetemplatesonline.com/">Free Web Templates</a> <a href="http://www.websitetemplates.org/">Website Templates</a> <a href="http://www.webdesign.org/website-design">Website Design</a></div>
     <div class="post">
@@ -19,12 +20,13 @@
             <h2 class="title">Modificar Modelo:</h2>
             <br/>
             
-            
-              ${error}
-              ${mensaje}
+             <script>
+             Mensaje("${mensaje}");
+             Mensaje("${error}")
+                </script>
         
 
-        <a href="eliminarModelo.htm?id=${p.codmodelo}" >Eliminar</a>
+        <a href="eliminarModelo.htm?id=${p.codmodelo}" onclick="return confirmar('¿Está seguro que desea eliminar este modelo?')">Eliminar</a>
         <br/></br>
         <form:form action="modificarModelo.htm?id=${p.codmodelo}" commandName="mm" >
             <table>
