+++ td/web/WEB-INF/jsp/modificarUsuario.jsp
@@ -12,6 +12,9 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@include file="../../interfaz/inicio.html" %>
 <!DOCTYPE html>
+
+<script type="text/javascript" src="interfaz/verificaciones.js"></script>
+   
 <div id="content"><div class="inner_copy"><a href="http://www.freetemplatesonline.com/">Free Web Templates</a> <a href="http://www.websitetemplates.org/">Website Templates</a> <a href="http://www.webdesign.org/website-design">Website Design</a></div>
     <div class="post">
 	
@@ -20,12 +23,15 @@
             
             <h2 class="title">Modificar Usuario:</h2>
             <br/>
+             <script>
+             Mensaje("${mensaje}");
+             Mensaje("${error}")
+                </script>
         
-             
-              ${error}
-              ${mensaje}
-
-         <a href="eliminarUsuario.htm?id=${p.codusuario}" >Eliminar</a> 
+           
+         <a href="eliminarUsuario.htm?id=${p.codusuario}" onclick="return confirmar('¿Está seguro que desea eliminar este usuario?')">Eliminar</a> 
+          
+         
          <br/><br/>
        
         <form:form action="modificarUsuario.htm?id=${p.codusuario}" commandName="mu" >
